+++ doc/tutorial/read_mapper/read_mapper_step4.cpp
@@ -136,7 +136,7 @@ void initialise_argument_parser(seqan3::argument_parser & parser, cmd_arguments 
                       seqan3::input_file_validator{{"index"}});
     parser.add_option(args.sam_path, 'o', "output", "The output SAM file path.",
                       seqan3::option_spec::DEFAULT,
-                      seqan3::output_file_validator{{"sam"}});
+                      seqan3::output_file_validator{seqan3::output_file_open_options::create_new, {"sam"}});
     parser.add_option(args.errors, 'e', "error", "Maximum allowed errors.",
                       seqan3::option_spec::DEFAULT,
                       seqan3::arithmetic_range_validator{0, 4});